this is my first serious game made by pygame
it contains enemy AI, shooting and health mechanics. Also includes the ability to crouch and also the enemies would crouch when they see you crouched so the game is a little bit more difficult and it is harder to dodge the bullets. With the current source code, all other options such as more powerful bullets with shorter ranges and this stuff can be added.